requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2.11
django-cors-headers==4.9.0
django-jazzmin==3.0.4
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
mysqlclient==2.2.8
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
Serializer==0.2.1
simplejson==3.20.2
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
